requirements.txt
PyMuPDF
scikit-image
Pillow
numpy
